mqtt5/fastapi/requirements.txt
paho-mqtt==1.5.0
uvicorn
fastapi
